Moved get_viewer_tools to glue.viewers.common so that it can be used by non-Qt viewers
