+++ glue/viewers/common/qt/data_viewer.py
@@ -12,6 +12,7 @@ from glue.viewers.common.qt.base_widget import BaseQtViewerWidget
 from glue.viewers.common.qt.tool import SimpleToolMenu
 from glue.viewers.common.qt.toolbar import BasicToolbar
 from glue.viewers.common.viewer import Viewer
+from glue.viewers.common.utils import get_viewer_tools
 
 __all__ = ['DataViewer', 'get_viewer_tools']
 
@@ -44,40 +45,6 @@ class SaveTool(SimpleToolMenu):
     icon = 'glue_filesave'
     tool_tip = 'Save/export the plot'
 
-
-def get_viewer_tools(cls, tools=None, subtools=None):
-    """
-    Given a viewer class, find all the tools and subtools to include in the
-    viewer.
-
-    Parameters
-    ----------
-    cls : type
-        The viewer class for which to look for tools.
-    tools : list
-        The list to add the tools to - this is modified in-place.
-    subtools : dict
-        The dictionary to add the subtools to - this is modified in-place.
-    """
-    if not issubclass(cls, DataViewer):
-        return
-    if tools is None:
-        tools = []
-    if subtools is None:
-        subtools = {}
-    if cls.inherit_tools and cls is not DataViewer:
-        for parent_cls in cls.__bases__:
-            get_viewer_tools(parent_cls, tools, subtools)
-    for tool_id in cls.tools:
-        if tool_id not in tools:
-            tools.append(tool_id)
-    for tool_id in cls.subtools:
-        if tool_id not in subtools:
-            subtools[tool_id] = []
-        for subtool_id in cls.subtools[tool_id]:
-            if subtool_id not in subtools[tool_id]:
-                subtools[tool_id].append(subtool_id)
-    return tools, subtools
 
 # Note: we need to use classmaker here because otherwise we run into issues when
 # trying to use the meta-class with the Qt class.
