+++ glue/viewers/common/qt/tests/test_data_viewer.py
@@ -108,43 +108,3 @@ class TestDataViewerImage(BaseTestDataViewer):
 
 class TestDataViewerHistogram(BaseTestDataViewer):
     widget_cls = HistogramViewer
-
-
-def test_get_viewer_tools():
-
-    class CustomViewer1(DataViewer):
-        pass
-
-    tools, subtools = get_viewer_tools(CustomViewer1)
-
-    assert tools == ['save']
-    assert subtools == {'save': []}
-
-    class CustomViewer2(DataViewer):
-        tools = ['banana']
-        subtools = {'save': ['apple', 'pear']}
-
-    tools, subtools = get_viewer_tools(CustomViewer2)
-
-    assert tools == ['save', 'banana']
-    assert subtools == {'save': ['apple', 'pear']}
-
-    CustomViewer2.inherit_tools = False
-
-    tools, subtools = get_viewer_tools(CustomViewer2)
-
-    assert tools == ['banana']
-    assert subtools == {'save': ['apple', 'pear']}
-
-    class Mixin(object):
-        pass
-
-    class CustomViewer3(CustomViewer2, Mixin):
-        tools = ['orange']
-        subtools = {'banana': ['one', 'two']}
-        inherit_tools = True
-
-    tools, subtools = get_viewer_tools(CustomViewer3)
-
-    assert tools == ['banana', 'orange']
-    assert subtools == {'save': ['apple', 'pear'], 'banana': ['one', 'two']}
